Updated ECommerce class to have defined global variables so all classes can share the connection to the database and the current user signed into the application
Updated HeaderPanel class with additional methods to handle the call to create the sign-in and new user registration form.

Updated SignInPanel with logic to parse the user information and validate a user for sign-in.

Added RegisterPanel class to display the new user registration form with logic to validate email address of new user and add new user to user.csv file.